pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "food_manager"
version = "0.0.1"
description = "REST API for food manager"
authors = [{name = "Célien Donzé", email = "celien.donze@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "SQLAlchemy",
    "pandas",
    "loguru",
]

[project.optional-dependencies]
test = ["pytest", "pytest-cov", "httpx"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["food_manager*"]


[tool.ruff]
select = ["F", "E", "D", "PD", "PL"]
ignore = ["D105", "E501"]

[tool.ruff.pydocstyle]
convention = "google"

# ignore errors for notebooks
[tool.ruff.extend-per-file-ignores]
"**/*.ipynb" = ["D100", "E402"]
"tests/*" = ["PLR2004"]